Seperate the scripts and updateDom files
